+++ pangeo_forge_recipes/recipes/base.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 
 from abc import ABC
-from dataclasses import dataclass, replace
+from dataclasses import dataclass, field, replace
 from typing import Callable, ClassVar
 
 from ..executors.base import Pipeline
 from ..patterns import FilePattern, prune_pattern
+from ..storage import StorageConfig, temporary_storage_config
 
 
 @dataclass
@@ -53,3 +54,21 @@ class FilePatternMixin:
 
         new_pattern = prune_pattern(self.file_pattern, nkeep=nkeep)
         return replace(self, file_pattern=new_pattern)
+
+
+@dataclass
+class StorageMixin:
+    """Provides the storage configuration for Pangeo Forge recipe classes.
+
+    :param storage_config: The storage configuration.
+    """
+
+    storage_config: StorageConfig = field(default_factory=temporary_storage_config)
+
+    @property
+    def target(self):
+        return f"{self.storage_config.target.fs.protocol}://{self.storage_config.target.root_path}"
+
+    @property
+    def target_mapper(self):
+        return self.storage_config.target.get_mapper()