+++ tests/recipe_tests/test_HDFReferenceRecipe.py
@@ -12,18 +12,24 @@ reference_hdf_zarr = pytest.importorskip("pangeo_forge_recipes.recipes.reference
 HDFReferenceRecipe = reference_hdf_zarr.HDFReferenceRecipe
 
 
+@pytest.mark.parametrize("default_storage", [True, False])
 @pytest.mark.parametrize("with_intake", [True, False])
-def test_single(netcdf_local_file_pattern_sequential, tmpdir, with_intake, execute_recipe):
+def test_single(
+    netcdf_local_file_pattern_sequential, tmpdir, with_intake, default_storage, execute_recipe,
+):
     file_pattern = netcdf_local_file_pattern_sequential
     path = list(file_pattern.items())[0][1]
     expected = xr.open_dataset(path, engine="h5netcdf")
     recipe = HDFReferenceRecipe(file_pattern)
 
-    # make sure assigning storage later works
-    out_target = FSSpecTarget(fs=fsspec.filesystem("file"), root_path=str(tmpdir))
-    metadata_cache = MetadataTarget(fs=fsspec.filesystem("file"), root_path=tempfile.mkdtemp())
-    recipe.target = out_target
-    recipe.metadata_cache = metadata_cache
+    if default_storage:
+        out_target = recipe.storage_config.target
+    else:
+        # make sure assigning storage later works
+        out_target = FSSpecTarget(fs=fsspec.filesystem("file"), root_path=str(tmpdir))
+        metadata_cache = MetadataTarget(fs=fsspec.filesystem("file"), root_path=tempfile.mkdtemp())
+        recipe.storage_config.target = out_target
+        recipe.storage_config.metadata = metadata_cache
 
     execute_recipe(recipe)
 
@@ -52,8 +58,8 @@ def test_multi(netcdf_local_file_pattern_sequential, tmpdir, with_intake, execut
     # make sure assigning storage later works
     out_target = FSSpecTarget(fs=fsspec.filesystem("file"), root_path=str(tmpdir))
     metadata_cache = MetadataTarget(fs=fsspec.filesystem("file"), root_path=tempfile.mkdtemp())
-    recipe.target = out_target
-    recipe.metadata_cache = metadata_cache
+    recipe.storage_config.target = out_target
+    recipe.storage_config.metadata = metadata_cache
 
     execute_recipe(recipe)